Fix boolean values in string column bug (#13)
## Description

Fix bug where boolean or null data in rows would not get rendered. This
is because false-y values are evaluated instead of rendered.

Also re-organized a lot of Table tests. The only new test is "renders
booleans and nulls in string columns"

Fixes PDS-2404

## Test plan

Write failing test and make it pass