Remove contextual ligatures in views (#19)
## Description

Remove contextual ligatures from views. Notably, this ensures that the `x` in `5x6` displays as a letter x rather than a multiplication symbol. This makes a view visually consistent with the rest of the Airplane UI.

## Test plan

I also added the inter font stylesheet to our example app so it matches a production view. Without this, I wasn't able to repro the ligature. With this new change, contextual ligatures do not show up.

---
FIXES PDS-2403