Improve chart loading state (#34)
## Description
Improve chart loading state so that it doesn't jump from a small bit of text to a full chart.

We display a loader the correct size as the chart if the user provides the size, or an approximate size if they don't. The loading state gracefully transitions to the chart via animation.

## Test plan
Check out the loading story in the storybook

Co-authored-by: None <48659329+aviator-app[bot]@users.noreply.github.com>