Use padding with tab panel instead of margin (#38)
## Description

Currently a button's focus ring is cut off on the top and on the side if the button is the first element in the tabs. This PR fixes one of those axes by using padding over margin. Padding doesn't cut off the border while margin does.

## Test plan
Previously both the left and top were cut off.
![image](https://user-images.githubusercontent.com/6363419/234992153-ccaeb7d8-19e5-49ef-a402-17b1143d0528.png)

---

CAP-2096

Co-authored-by: None <48659329+aviator-app[bot]@users.noreply.github.com>