Reverse text preset and heading preset (#46)
## Description
Line up the heading and text presets with heading order, i.e. the smallest number is the largest.

As we currently have it, we're passing in a reversed (incorrect) list as the `headings` theme. We could reverse it in `theme.ts`, but I think it's easier to do what I did here.

## Test plan
I'm relying on chromatic to point out any UI diffs

fixes https://github.com/airplanedev/views/issues/45