Grow chart to 100% of container (#53)
## Description
Grow chart to fill 100% of container. This is similar to how table works where it will fill as much space as it needs and shrink down as appropriate.

If I place two charts next to each other in a stack
Before
![image](https://github.com/airplanedev/views/assets/6363419/1f9661be-0816-41a4-b1fe-d1021eef6932)
After
![image](https://github.com/airplanedev/views/assets/6363419/caba19c6-7797-4e05-8cd3-e791bed494b8)

One chart in a stack
Before
![image](https://github.com/airplanedev/views/assets/6363419/a39f83be-cd36-4932-a7f3-d574c3796d17)
After
![image](https://github.com/airplanedev/views/assets/6363419/dc0d6385-396e-45d8-b15a-fbf836c8d160)


## Test plan
I played around with a bunch of chart examples