Explicitly set chart height (#56)
## Description

Explicitly set chart height rather than relying on the height set by the underlying library.

We thought that the library might be setting the height intelligently based on the data, but it's actually just hardcoding it to 450px. We'd rather control the height so we can decide what the ideal height is (375 seemed a bit better to view the data and view multiple charts on the same page) and use this height for our loading and error states.

## Test plan
Our chart storybooks will update and I added a new storybook to confirm that you can customize the height.