Fix form FileInputs and release v2.1.2 (#65)
## Description

If a `FileInput` submits only a single file to a `Form` in order to run a task, the form would ignore it and return `undefined` for the parameter instead. This was affecting both task-backed forms with `upload` parameters and normal `FileInput` components.

Context from Airplane Slack: https://airplanedev.slack.com/archives/C0374J5B745/p1686777030198279

## Test plan

Create a task-backed `Form` for a task with a required `upload` parameter. Confirm that the upload parameter isn't submitted to the task before this change, but is correctly sent with this change.