Support react 17 by removing useid (#73)
## Description

`useID` was added in React 18, so we have to remove it to support React 16-17.

In addition, I had to use a version of the dialog that is not connected to the global component state within a table confirmation button. Unsure why the change to React 17 necessitates this, but this removes an infinite loop.

On that note, we should generally be using non-connected components when we use components internally as there is some overhead to using connected components.


## Test plan
Manually tested that I can render a table using React 17.