Refresh task-backed table after row action menu completes (#79)
## Description

We have in our docs that "When row action in a task backed table executes successfully, the table automatically refreshes, calling its backing task. This ensures that the table is up to date even if the row action mutates data." This is true of `rowActions` but we never implemented this for `rowActionsMenu`.

Resolves CAP-2957

## Test plan

yarn build and then yarn add local views path. Verify in repro files included in ticket that row action menu refreshes the task-backed query.