Update js-sdk version and set useSelfHostedInputs  (#97)
## Description
This change updates views to use the latest js-sdk version so we can pick up self-hosted inputs support. It also sets the `useSelfHostedInputs` client option when executing tasks so that self-hosted inputs are enabled if the view's team supports them.

## Test plan
Testing completed successfully via running the `adminPanel` example locally and verifying that run inputs were saved in the correct place.