Export marker that says when a task is executing (#100)
## Description

Correctly document the behavior of `useTaskQuery` `loading` as when the task is first executed. Create a new `executing` field that is true any time the task is executed which includes the first time and any additional times the task is revalidated.

## Test plan
Manually tested and added some unit tests

Manual test confirming that both `useRefetchTasks` and the `refetch` callback returned by `useTaskQuery` trigger the `executing` state. 
<img width="637" alt="image" src="https://github.com/airplanedev/views/assets/6363419/cbcbeb9e-f76b-4b70-ae34-fe2550218ff5">

FIXES CAP-3171