Support task backed select in views task backed form (#103)
## Description

Port over changes supporting task backed select to views. Most of the logic here is duplicated from the main app, but the code is organized differently so while the spirit is the same, most of the code was written from scratch.

We also were using a lot of incorrect types for parameters. I made some fixes here, but it still isn't perfect and there are some places we are forcing things into strings e.g.

## Test plan
Wrote unit test for core functionality. Manually tested with a task-backed select