Fix JST identification (#112)
## Description

Fix what we identify as a JST in view params. There was a point when we were incorrectly storing JSTs in options as template objects. The views solution was developed during this time. JSTs are actually stored as strings, so update this logic to exactly match the logic in web: https://github.com/airplanedev/web/blob/1926779bdad15001c0dda1f96994df3af68d187b/components/parameters/ParameterInput.tsx#L627

Then I updated the template evaluation hook to echo back values that don't contain templates which required copying over the template parsing from web.

## Test plan
I added a test that ensures that a string template is evaluated and a string non-template is not.