Remove div wrapper on task backed button tooltips (#114)
## Description
Task-backed buttons currently have tooltips with div wrappers. The issue is that the div wrapper prevents width and height props from being properly applied to task-backed buttons. In removing the div wrapper, we have to add pointerEvents: all so that the tooltip text shows up.

## Test plan
Chromatic shows no regressions.

Tested a view with the following components. Button is cut off before the change, and takes up 2/3 of the screen width after the change.
```
<Stack direction="row" width="full">
  <Button task="list_accounts" width="2/3">
    Click Me
  </Button>
</Stack>
```