Support task caching for views queries (#133)
## Description
This change updates views to support server-based task caching. This is done via a new query option, `allowCachedMaxAge`, which, if set, is passed to the execution flow in the js-sdk. This change also sets `isCached` and `cacheFetchedAt` in run started messages to the web UI so that we can properly show whether runs are cached in the view activity panel.

## Test plan
Testing completed successfully via running `adminPanel` example locally and verifying that `allowCachedMaxAge` was set correctly.